requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy[asyncio]==2.0.36
aiomysql==0.2.0
aiosqlite==0.20.0
python-dotenv==1.0.1
jinja2==3.1.4
httpx==0.28.1
pyyaml==6.0.2
python-multipart==0.0.18
cryptography==44.0.0
